pyproject.toml
[project]
name = "gen-ai-gemini-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.2.1",
]
